+++ tests/tools/test_add_year.py
@@ -1,10 +1,16 @@
-# -*- coding: utf-8 -*-
-from message_ix import Scenario
-from message_ix.tools.add_year import add_year
 import pytest
 
+from ixmp import Platform
+from message_ix import Scenario
+from message_ix.tools.add_year import add_year
 
-def model_setup(scen, data):
+
+@pytest.fixture
+def base_scen_mp(test_mp):
+    scen = Scenario(test_mp, 'model', 'standard', version='new')
+
+    data = {2020: 1, 2030: 2, 2040: 3}
+
     years = sorted(list(set(data.keys())))
     scen.add_set('node', 'node')
     scen.add_set('commodity', 'comm')
@@ -20,6 +26,11 @@ def model_setup(scen, data):
         tec_specs = ['node', 'tec', yr, yr, 'mode']
         scen.add_par('output', tec_specs + output_specs, 1, '-')
         scen.add_par('var_cost', tec_specs + ['year'], value, 'USD/GWa')
+
+    scen.commit('initialize test model')
+    scen.solve(case='original_years')
+
+    yield scen, test_mp
 
 
 def adding_years(test_mp, scen_ref, years_new):
@@ -63,16 +74,60 @@ def assert_function(scen_ref, scen_new, years_new, yr_test):
     assert value_new == pytest.approx(value_ref, rel=1e-04)
 
 
-def test_add_year(test_mp):
-    scen_ref = Scenario(test_mp, 'model', 'standard', version='new')
-    model_setup(scen_ref, {2020: 1, 2030: 2, 2040: 3})
-    scen_ref.commit('initialize test model')
-    scen_ref.solve(case='original_years')
+YEARS_NEW = [2025, 2035]
+
+
+def test_add_year(base_scen_mp):
+    scen_ref, test_mp = base_scen_mp
 
     # Adding new years
-    years_new = [2025, 2035]
-    scen_new = adding_years(test_mp, scen_ref, years_new)
+    scen_new = Scenario(test_mp, model='add_year', scenario='standard',
+                        version='new', annotation=' ')
+    add_year(scen_ref, scen_new, YEARS_NEW)
     scen_new.solve(case='new_years')
 
     # Running the tests
-    assert_function(scen_ref, scen_new, years_new, yr_test=2025)
+    assert_function(scen_ref, scen_new, YEARS_NEW, yr_test=2025)
+
+
+def test_add_year_cli(message_ix_cli, base_scen_mp):
+    scen_ref, test_mp = base_scen_mp
+
+    # Information about the base Scenario
+    platform_name = test_mp.name
+    model = scen_ref.model
+    scenario = scen_ref.scenario
+
+    cmd = [
+        '--platform', platform_name,
+        '--model', model,
+        '--scenario', scenario,
+        'add-years',
+        '--years_new', repr(YEARS_NEW),
+        '--model_new', 'add_year',
+        '--scen_new', 'standard',
+    ]
+
+    # Delete the objects so that the database connection is closed
+    del test_mp, scen_ref
+
+    r = message_ix_cli(*cmd)
+    print(r.output, r.exception)
+    assert r.exit_code == 0
+
+    # Re-load the base Scenario
+    mp = Platform(name=platform_name)
+    scen_ref = Scenario(mp, model=model, scenario=scenario)
+
+    # Load the created Scenario
+    scen_new = Scenario(mp, model='add_year', scenario='standard')
+
+    assert_function(scen_ref, scen_new, YEARS_NEW, yr_test=2025)
+
+    # Same, except with --dry-run
+    r = message_ix_cli(*cmd, '--dry-run')
+    assert r.exit_code == 0
+
+    # Bad usage: not giving the base scenario info
+    r = message_ix_cli(*cmd[6:], '--dry-run')
+    assert r.exit_code == 2