+++ RELEASE_NOTES.md
@@ -12,7 +12,8 @@
 
 - [#256](https://github.com/iiasa/message_ix/pull/256): Update to use :obj:`ixmp.config` and improve CLI.
 - [#255](https://github.com/iiasa/message_ix/pull/249): Add :mod:`message_ix.testing.nightly` and `nightly` CLI command group for slow-running tests.
-- [#249](https://github.com/iiasa/message_ix/pull/249): Build MESSAGE and MESSAGE_MACRO classes on ixmp model API; adjust Scenario.
+- [#249](https://github.com/iiasa/message_ix/pull/249),
+  [#259](https://github.com/iiasa/message_ix/pull/259): Build MESSAGE and MESSAGE_MACRO classes on ixmp model API; adjust Scenario.
 - [#235](https://github.com/iiasa/message_ix/pull/236): Add a reporting tutorial.
 - [#236](https://github.com/iiasa/message_ix/pull/236),
   [#242](https://github.com/iiasa/message_ix/pull/242): Enhance reporting.